made cleanup forcable in config, ensured it's only ran once

+++ tasks-available/self-update.py
@@ -22,6 +22,7 @@ import mte_task_dispatcher
 class Task(mte_task_dispatcher.Task):
   def __init__(self, core, task_name):
     # Run template task initiator
+    self.clean = False
     super().__init__(core, task_name) 
 
   # Everything that needs to be executed should be called from the
@@ -30,11 +31,15 @@ class Task(mte_task_dispatcher.Task):
     self.core.log.add('[' + self.get_task_name() + '] Getting updates from git.', 5)
     for line in self.core.run_command('git pull', self.get_task_name()):
       self.core.log.add('[' + self.get_task_name() + '] ' + line, 4)
+    if self.core.config.get('always_cleanup', self.get_task_name()):
+      self.cleanup()
   
   def cleanup(self):
-    self.core.log.add('[' + self.get_task_name() + '] Running cleanup', 5)
-    # Clear cache/
-    for line in self.core.run_command('rm -rf cache/*', self.get_task_name()):
-      if len(line) > 0:
-        self.core.log.add('[' + self.get_task_name() + ']' + line, 4)
+    if self.clean == False:
+      self.core.log.add('[' + self.get_task_name() + '] Running cleanup', 5)
+      # Clear cache/
+      for line in self.core.run_command('rm -rf cache/*', self.get_task_name()):
+        if len(line) > 0:
+          self.core.log.add('[' + self.get_task_name() + ']' + line, 4)
+      self.clean = True
